requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
flask_cors==4.0.1
flask_migrate==4.0.7
flask_restful==0.3.10
flask_swagger_ui==4.11.1
gunicorn==22.0.0
pydantic==2.7.1
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.19
